% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binary_search_cpp}
\alias{binary_search_cpp}
\title{Binary search cpp version}
\usage{
binary_search_cpp(x, y)
}
\arguments{
\item{x}{A double to be find}

\item{y}{A double vector}
}
\value{
A integer of position
}
\description{
Binary search cpp version
}
\author{
Zhaowei
}
