% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binary_search_r}
\alias{binary_search_r}
\title{Get the position of x in vector y}
\usage{
binary_search_r(x, y, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{A numeric value}

\item{y}{A numeric vector}

\item{tol}{double equal tolerance}
}
\value{
A integer of the \code{x} position in \code{y}.
}
\description{
Get the position of x in vector y
}
\examples{
binary_search_r(2.5, 1:10)
binary_search_r(1.2, 1:10)
binary_search_r(0.5, 1:10)
binary_search_r(9.5, 1:10)
binary_search_r(11, 1:10)

}
\author{
\href{https://github.com/BruceZhaoR}{Wei Zhao}
}
