% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_utils.R
\name{cal_gps_dist}
\alias{cal_gps_dist}
\title{Calculate earth distance between two points gps longitude and latitude.}
\usage{
cal_gps_dist(lon_s, lat_s, lon_e, lat_e)
}
\arguments{
\item{lon_s}{Start gps longitude}

\item{lat_s}{Start gps latitude}

\item{lon_e}{End gps longitude}

\item{lat_e}{End gps latitude}
}
\value{
Distance of double type in meters.
}
\description{
Calculate earth distance between two points gps longitude and latitude.
}
\examples{
cal_gps_dist(121.434174,31.158213,121.444366,31.169403)
# meters
}
