% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_user_coef.R
\name{cal_user_coef}
\alias{cal_user_coef}
\title{A big wrapper of get user insurance discount coef.}
\usage{
cal_user_coef(df, risk_names, score_config_list, rf_weight, discount_index)
}
\arguments{
\item{df}{Raw risk factors data.frame/tibble with the correct col names.}

\item{risk_names}{Correct col names to calculate.}

\item{score_config_list}{A List contains all riskFactors value and score
index, read from .RDS file.}

\item{rf_weight}{Correct col names of risk factors weight, sum to 1.}

\item{discount_index}{Specified score and coef.}
}
\value{
a \code{\link[tibble:as_tibble]{tibble}} / data.frame
}
\description{
This function wrap all the functions to get the raw risk factors speicified
score, daymileage, discount coef. Before use this function, pleage make sure
that your data and cofiguration are correct.
}
\examples{
\dontrun{
cal_user_coef(pj_rf, "mielage", score_config_list, rf_weight, discount_index)

risk_names <- c("acc_count_phk","act_radius","dec_count_phk","high_curv_tr","holiday_tr",
             "interstate_r","lane_change_phk","late_night_tr","long_tr","main_act_prov" ,
             "mileage","speeding_lvl", "speeding_phk", "trip_dis_e", "turn_count_phk")
cal_user_coef(pj_rf, risk_names, score_config_list, rf_weight, discount_index)
}
}
\author{
\href{https://github.com/BruceZhaoR}{Wei Zhao}
}
