% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_utils.R
\name{coordinate_transform}
\alias{coordinate_transform}
\alias{wgs2gcj}
\alias{gcj2wgs}
\alias{gcj2bd}
\alias{bd2gcj}
\alias{wgs2bd}
\alias{bd2wgs}
\title{Coordinate transform}
\usage{
wgs2gcj(wgs_lon, wgs_lat)

gcj2wgs(gcj_lon, gcj_lat)

gcj2bd(gcj_lon, gcj_lat)

bd2gcj(bd_lon, bd_lat)

wgs2bd(wgs_lon, wgs_lat)

bd2wgs(bd_lon, bd_lat)
}
\arguments{
\item{wgs_lon}{longitude}

\item{wgs_lat}{latitude}

\item{gcj_lon}{longitude}

\item{gcj_lat}{latitude}

\item{bd_lon}{longitude}

\item{bd_lat}{latitude}
}
\value{
Transformed vector of  longitude and latitude.
}
\description{
WGS84 to GCJ02, GCJ02 to BD09 and vice versa.
WG84 is used by Google maps and hardware,
GCJ02 is used by Gaode and Tencent,
BD09 is used by Baidu.
}
\examples{
test <-
  structure(
    list(
      startLon = c(
        113.58850000000001,
        113.54700000000003,
        113.54050000000001,
        113.52866660000001,
        113.53083329999998,
        113.54083329999997
      ),
      startLat = c(
        33.43299999999999,
        33.466666599999996,
        33.49250000000001,
        33.4581666,
        33.45766660000001,
        33.49250000000001
      ),
      endLon = c(
        113.5468333,
        113.54050000000001,
        113.52866660000001,
        113.53083329999998,
        113.54083329999997,
        113.52249999999998
      ),
      endLat = c(
        33.465,
        33.49250000000001,
        33.4581666,
        33.457166599999994,
        33.49233330000001,
        33.49250000000001
      )
    ),
    row.names = c(NA, -6L),
    class = c("tbl_df", "tbl", "data.frame")
  )

wgs_lon <- test$startLon[[1]]
wgs_lat <- test$startLat[[1]]
wgs2gcj(wgs_lon, wgs_lat)# gaode
wgs2bd(wgs_lon, wgs_lat)#baidu

t(apply(test, 1, function(x) {
  start_gps <- wgs2bd(x[1], x[2])
  end_gps <- wgs2bd(x[3], x[4])
  c(start_gps[1], start_gps[2], end_gps[1], end_gps[2])
}))

}
