% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_utils.R
\name{fmt_gps_json}
\alias{fmt_gps_json}
\title{Generate GPS lines data for echarts}
\usage{
fmt_gps_json(df)
}
\arguments{
\item{df}{data.frame with starLon, starLat, endLon, endLat.}
}
\value{
json txt
}
\description{
Your can see the package/inst/misc files.
}
\examples{
\dontrun{
test <- select(df, startLon, startLat, endLon, endLat)
as_tibble(t(apply(test, 1, function(x) {
  x <- as.numeric(x)
  start_gps <- wgs2bd(x[1], x[2])
  end_gps <- wgs2bd(x[3], x[4])
  c(start_gps[1], start_gps[2], end_gps[1], end_gps[2])
})))
}
}
