% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScore.R
\name{get_score}
\alias{get_score}
\alias{get_score_vec}
\title{Get the score of a risk factor according to the given config.}
\usage{
get_score(riskfactor, values, scores)

get_score_vec(rf_vec, values, scores)
}
\arguments{
\item{riskfactor}{Raw risk factor.}

\item{values}{Specified risk factors.}

\item{scores}{Specified score according to the value.}

\item{rf_vec}{A vector of risk factors}
}
\value{
A numeric score.

A vector of score
}
\description{
Use lapply to a vector of risk factors and return a vector of scores.
}
\examples{
values <- c(55, 60, 65, 70, 75, 80, 85, 90, 95, 100)
scores <- c(1.0000, 0.9747, 0.8967, 0.8303, 0.7730, 0.7232, 0.6793, 0.6405, 0.6059, 0.5748)
get_score(57.5, values, scores) # 0.98735
get_score(97.5, values, scores) # 0.59035
get_score(77, values, scores) # 0.75308

get_score_vec(c(57.5, 97.5,77), values, scores)

}
\author{
\href{https://github.com/BruceZhaoR}{Wei Zhao}
}
