% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjutils-package.R
\docType{package}
\name{pjutils}
\alias{pjutils}
\alias{pjutils-package}
\title{pjutils: Utils for data calculation and analysis.}
\description{
This package wraps the most common used functions for calculating risk factors
to score and some other userful functions.
}
