% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-tibble.r, R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{data_frame}
\alias{as_data_frame}
\alias{add_row}
\alias{glimpse}
\alias{tibble}
\alias{as_tibble}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{data_frame}}, \code{\link[tibble]{as_data_frame}}, \code{\link[tibble]{add_row}}, \code{\link[tibble]{glimpse}}, \code{\link[tibble]{tibble}}, \code{\link[tibble]{as_tibble}}}
}}

