% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScore.R
\name{rf2score}
\alias{rf2score}
\title{Raw risk factors value to score.}
\usage{
rf2score(rf_df, score_config_list)
}
\arguments{
\item{rf_df}{Risk factors data frame(exclude user_id)}

\item{score_config_list}{A List contains all riskFactors value and score
index, read from .RDS file.}
}
\value{
a \code{\link[tibble:as_tibble]{tibble}} with every risk factor score
}
\description{
The input risk factors should be a data.frame or a
\code{\link[tibble:as_tibble]{tibble}}, which contains all raw risk factors
value and should not contians user_id and others. The
\code{score_config_list} should be a list, \code{*_values} should be in front
of \code{*_scores}.
}
\details{
You should carefull about the risk facotrs \code{rf_df} names and the
\code{score_config_list} names. \code{*_values}/\code{*_scores} and the risk
factors names should be the same.
}
\examples{
\dontrun{
require(dplyr)
score_config_list <- readRDS("result/score_config_list.RDS")
pj_rf <- readr::read_csv("data/pingjia/pingjia_result.csv",col_names = FALSE)
nameTmp <- c("acc_count_phk","act_radius","dec_count_phk","high_curv_tr","holiday_tr",
"interstate_r","lane_change_phk","late_night_tr","long_tr","main_act_prov"
, "mileage","speeding_lvl", "speeding_phk", "trip_dis_e", "turn_count_phk","user_id")
names(pj_rf) <- nameTmp

risk_factor_names <- setdiff(nameTmp, c("user_id","main_act_prov"))
pj_rf <- pj_rf \%>\% select(!!risk_factor_names)

rf2score(pj_rf, score_config_list)
}

}
\author{
ZhaoWei
}
